requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
cryptography==3.4.7
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==2.0.1
multidict==5.1.0
mysql-connector==2.2.9
Naked==0.1.31
pycparser==2.20
pycrypto==2.6.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==5.4.1
qiniu==7.4.1
redis==3.5.3
requests==2.26.0
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.22
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.5.0
